정산 상태 변경 테스트 완료

+++ src/main/java/backend/codebackend/controller/BasketController.java
@@ -4,13 +4,16 @@ import backend.codebackend.domain.Basket;
 import backend.codebackend.domain.Member;
 import backend.codebackend.service.BasketService;
 import backend.codebackend.service.MemberService;
+import backend.codebackend.service.MozipService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -21,27 +24,40 @@ import java.util.Optional;
 public class BasketController {
     private final BasketService basketService;
     private final MemberService memberService;
+    private final MozipService mozipService;
     @PostMapping("/basket")
     public List<Basket> listBasketItems(@PathVariable Long chatroom_id) {
         List<Basket> basketItems = basketService.findAll(chatroom_id);
         return basketItems;
     }
 
+
     @PostMapping("/basket/add")
-    public Basket addItemToBasket(Long chatroom_id, String menuName, String menuPrice, HttpServletRequest request) {
+    public ResponseEntity<?> addItemToBasket(Long chatroom_id, String menuName, String menuPrice, HttpServletRequest request) {
         HttpSession session = request.getSession(false);
         if (session == null) {
             log.info("세션이 없습니다.");
-            return null;
+            return ResponseEntity.badRequest().build();
+        }
+
+        //정산 상태 확인
+        if(mozipService.mozipStatus(chatroom_id)) {
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "정산 시작된 상태입니다!"));
         }
 
         Optional<Member> member = memberService.findLoginId(String.valueOf(session.getAttribute("memberId")));
         //menuPrice : "10,000원" 에서 숫자 이외의 모든 문자 제거 후 int 타입 변수에 대입
         //value : member 값이 null이 아닐 경우 객체, orElse : member 값이 null인 경우 null
-        return member.map(value -> basketService.addItemToBasket(chatroom_id, menuName
+        Basket basket = member.map(value -> basketService.addItemToBasket(chatroom_id, menuName
                 , Integer.parseInt(menuPrice.replaceAll("[^0-9]", ""))
                 , value.getNickname()
         )).orElse(null);
+
+        if (basket == null) {
+            return null;
+        } else {
+            return ResponseEntity.ok(basket);
+        }
     }
 
     //다른 사람이 추가한 최신 메뉴를 받음 (STOMP)
@@ -69,55 +85,77 @@ public class BasketController {
     //증가 버튼
     @PostMapping("/chat/basket/plusQuantity")
     @ResponseBody
-    public String plusQuantity(Long menuId, HttpServletRequest request) {
+    public ResponseEntity<?>  plusQuantity(Long chatroom_id, Long menuId, HttpServletRequest request) {
         HttpSession session = request.getSession(false);
+        
+        if (session == null) {
+            log.info("세션이 없습니다.");
+            return ResponseEntity.badRequest().build();
+        }
+        
+        //정산 상태 확인
+        if(mozipService.mozipStatus(chatroom_id)) {
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "정산 시작된 상태입니다!"));
+        }
 
         //HTML의 menuId를 수정하여 접근하는 유저를 방지하기 위해 장바구니 실제로 담은 사람이랑 로그인 유저랑 비교한다.
         String nickname = memberService.findLoginId(String.valueOf(session.getAttribute("memberId"))).get().getNickname();
 
         if(!basketService.findBasketMenu(menuId).getNickname().equals(nickname)) {
-            return null;
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "본인의 메뉴만 수정 할 수 있습니다."));
         }
 
         basketService.plusItemCnt(menuId, nickname);
 
-        return "+ 수량 수정";
+        return ResponseEntity.ok("수량 수정");
     }
 
     //감소 버튼
     @PostMapping("/chat/basket/minusQuantity")
     @ResponseBody
-    public String minusQuantity(Long menuId, HttpServletRequest request) {
+    public ResponseEntity<?>  minusQuantity(Long chatroom_id, Long menuId, HttpServletRequest request) {
         HttpSession session = request.getSession(false);
+
+        //정산 상태 확인
+        if(mozipService.mozipStatus(chatroom_id)) {
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "정산 시작된 상태입니다!"));
+        }
 
         //HTML의 menuId를 수정하여 접근하는 유저를 방지하기 위해 장바구니 실제로 담은 사람이랑 로그인 유저랑 비교한다.
         String nickname = memberService.findLoginId(String.valueOf(session.getAttribute("memberId"))).get().getNickname();
 
         if(!basketService.findBasketMenu(menuId).getNickname().equals(nickname)) {
-            return null;
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "본인의 메뉴만 수정 할 수 있습니다."));
         }
 
         basketService.minusItemCnt(menuId, nickname);
-        return "- 수량 수정";
+        return  ResponseEntity.ok("- 수량 수정");
     }
 
     //메뉴 삭제 버튼
     @PostMapping("/chat/basket/deleteByMenu")
     @ResponseBody
-    public String deleteByMenu(Long menuId, HttpServletRequest request) {
+    public ResponseEntity<?> deleteByMenu(Long chatroom_id, Long menuId, HttpServletRequest request) {
         HttpSession session = request.getSession(false);
 
-        if(session == null)
-            return null;
+        if (session == null) {
+            log.info("세션이 없습니다.");
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "세션이 없습니다. 다시 로그인 해주세요."));
+        }
 
+        //정산 상태 확인
+        if(mozipService.mozipStatus(chatroom_id)) {
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "정산 시작된 상태입니다!"));
+        }
+        
         //HTML의 menuId를 수정하여 접근하는 유저를 방지하기 위해 장바구니 실제로 담은 사람이랑 로그인 유저랑 비교한다.
         String nickname = memberService.findLoginId(String.valueOf(session.getAttribute("memberId"))).get().getNickname();
         if(!basketService.findBasketMenu(menuId).getNickname().equals(nickname)) {
-            return null;
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "본인의 메뉴만 삭제 할 수 있습니다."));
         }
 
         basketService.deleteByMenu(menuId);
-        return "장바구니 메뉴 삭제";
+        return ResponseEntity.ok("장바구니 삭제");
     }
 
     //장바구니 총 금액 뷰
